[fix][chaejy]: post type api별로 구분 및 type명 수정
