+++ src/entities/post/model/types.ts
@@ -1,30 +1,11 @@
 import type { ImageItem } from '@/entities/post/api/types';
 
-export type PostCategory = 'all' | 'event' | 'activity' | 'partnership' | 'release' | 'others';
+export type PostCategory = 'all' | 'event' | 'activity' | 'partnership' | 'patch' | 'etc';
 
 export type CategoryBadge = Exclude<PostCategory, 'all'>;
 
-export const POST_CATEGORY_LABEL_MAP: Record<PostCategory, string> = {
-  all: '전체',
-  event: '행사',
-  activity: '활동',
-  partnership: '제휴',
-  release: '패치',
-  others: '기타',
-};
-
-export const CATEGORY_BADGE_LABEL_MAP: Record<CategoryBadge, string> = {
-  event: '행사',
-  activity: '활동',
-  partnership: '제휴',
-  release: '패치',
-  others: '기타',
-};
-
-export const RESERVATION_LABEL = '예약중';
-
 export type Post = {
-  id: number;
+  postId: number;
   title: string;
   content: string;
   writer: string;
