+++ src/entities/post/ui/post-card/PostCard.stories.tsx
@@ -15,16 +15,20 @@ export default meta;
 type Story = StoryObj<typeof PostCard>;
 
 const mockPost: Post = {
-  id: 1,
+  postId: 1,
   title: '프론트엔드 스터디 모집합니다!',
   content: 'React 19, Zustand, Tanstack Query 중심으로 학습해요 🚀',
-  writer: '보라',
+  writer: '김테비',
   date: '2025-11-12',
+  pinned: false,
+  isReserved: false,
+  boardId: 1,
   likeCount: 42,
   isLiked: false,
+  scrappedByMe: false,
+  scrapCount: 0,
   commentCount: 8,
   category: 'event',
-  isReserved: false,
   thumbnailUrl: 'https://images.unsplash.com/photo-1522204507765-4b9e8f69e4f1?w=200&h=200&fit=crop',
 };
 
@@ -54,7 +58,7 @@ export const AdminView: Story = {
       ...mockPost,
       isReserved: true,
     },
-    currentCategory: 'release',
+    currentCategory: 'patch',
     userLevel: 'manager',
   },
 };
